fix(sync): publish relay state changes and make the spawn claim atomic
Addresses all four review findings from CodeRabbit and Codex on #949.

Publish relay suppression transitions (Codex). A running tunnel that took
a 4505 updated controlSuppressedReason without emitting a sync-status
snapshot, and the desktop only seeds once and then listens. On an idle
machine the new banner therefore stayed hidden until some unrelated sync
activity happened along — exactly the silence this change exists to end.
Entering and leaving suppression now publish route state, and syncService
grew notifyRouteStateChanged so the machine-level tunnel can nudge a
project-scoped emitter.

Treat routed sync events as invalidation, not payload (Codex). On a
remote-bound project the preload subscription fans out the remote
runtime's snapshot too, so applying it directly let a remote outage raise
a warning about this machine, or let remote health mask this machine's
own. AppShell now re-reads getLocalStatus on every event, matching
useSyncConnections.

Make the runtime spawn claim atomic (Codex). Two CLI processes reaching
an absent record concurrently both passed the check and both launched a
detached brain — the burst the record exists to prevent, and on named-pipe
platforms the loser cannot even use the socket-liveness abort. The TUI
already had the right primitive, so withSocketSpawnLock moves to
services/runtime/socketSpawnLock.ts and now wraps check-and-spawn on both
paths. Three independent reviewers flagged this duplication; it is now one
mechanism.

Serialize tunnel lifecycle transitions (CodeRabbit). stop() and start()
are async, so a lease reacquired mid-teardown could be shut down by the
stop already in flight. A generation token guards both, and a stop that
settles after a newer start re-establishes the tunnel instead of leaving
it down.
